Start from scratch
Revert "Merge pull request #10 from apollographql/course-steps-v2"

b5c5721b3505fb07499c1e9bb237e81d6a75d6f0
