+++ final/package.json
@@ -5,8 +5,6 @@
   "main": "index.js",
   "scripts": {
     "start": "nodemon index.js",
-    "start:monolith-subgraph": "nodemon monolith.js",
-    "compose": "rover supergraph compose --config supergraph-config.yaml > supergraph.graphql",
     "launch": "node utils/launch.js",
     "postinstall": "npm run install-services",
     "install-services": "node utils/install.js",
@@ -19,8 +17,6 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@apollo/gateway": "^0.50.0",
-    "@apollo/subgraph": "^0.1.5",
     "@prisma/client": "^3.0.2",
     "apollo-datasource": "^3.1.0",
     "apollo-datasource-rest": "^3.2.0",
@@ -28,7 +24,7 @@
     "axios": "^0.24.0",
     "concurrently": "^6.3.0",
     "date-fns": "^2.25.0",
-    "graphql": "^15.8.0",
+    "graphql": "^15.5.3",
     "nodemon": "^2.0.12",
     "sequelize": "^6.9.0",
     "sqlite3": "^5.0.2",
