+++ final/resolvers.js
@@ -3,6 +3,53 @@ const authErrMessage = '*** you must be logged in ***';
 
 const resolvers = {
   Query: {
+    user: async (_, { id }, { dataSources }) => {
+      const user = await dataSources.accountsAPI.getUser(id);
+      if (!user) {
+        throw new Error('No user found for this Id');
+      }
+      return user;
+    },
+    me: async (_, __, { dataSources, userId }) => {
+      if (!userId) throw new AuthenticationError(authErrMessage);
+      const user = await dataSources.accountsAPI.getUser(userId);
+      return user;
+    },
+    searchListings: async (_, { criteria }, { dataSources }) => {
+      const { numOfBeds, checkInDate, checkOutDate, page, limit, sortBy } = criteria;
+      const listings = await dataSources.listingsAPI.getListings({ numOfBeds, page, limit, sortBy });
+
+      // check availability for each listing
+      const listingAvailability = await Promise.all(
+        listings.map((listing) =>
+          dataSources.bookingsDb.isListingAvailable({ listingId: listing.id, checkInDate, checkOutDate })
+        )
+      );
+
+      // filter listings data based on availability
+      const availableListings = listings.filter((listing, index) => listingAvailability[index]);
+
+      return availableListings;
+    },
+    hostListings: async (_, __, { dataSources, userId, userRole }) => {
+      if (!userId) throw new AuthenticationError(authErrMessage);
+
+      if (userRole === 'Host') {
+        return dataSources.listingsAPI.getListingsForUser(userId);
+      } else {
+        throw new ForbiddenError('Only hosts have access to listings.');
+      }
+    },
+    listing: (_, { id }, { dataSources }) => {
+      return dataSources.listingsAPI.getListing(id);
+    },
+    featuredListings: (_, __, { dataSources }) => {
+      const limit = 3;
+      return dataSources.listingsAPI.getFeaturedListings(limit);
+    },
+    listingAmenities: (_, __, { dataSources }) => {
+      return dataSources.listingsAPI.getAllAmenities();
+    },
     guestBookings: async (_, __, { dataSources, userId, userRole }) => {
       if (!userId) throw new AuthenticationError(authErrMessage);
 
@@ -51,6 +98,24 @@ const resolvers = {
     },
   },
   Mutation: {
+    updateProfile: async (_, { updateProfileInput }, { dataSources, userId }) => {
+      if (!userId) throw new AuthenticationError(authErrMessage);
+      try {
+        const updatedUser = await dataSources.accountsAPI.updateUser({ userId, userInfo: updateProfileInput });
+        return {
+          code: 200,
+          success: true,
+          message: 'Profile successfully updated!',
+          user: updatedUser,
+        };
+      } catch (err) {
+        return {
+          code: 400,
+          success: false,
+          message: err.message,
+        };
+      }
+    },
     createBooking: async (_, { createBookingInput }, { dataSources, userId }) => {
       if (!userId) throw new AuthenticationError(authErrMessage);
 
@@ -81,6 +146,65 @@ const resolvers = {
           success: true,
           message: 'Successfully booked!',
           booking,
+        };
+      } catch (err) {
+        return {
+          code: 400,
+          success: false,
+          message: err.message,
+        };
+      }
+    },
+    createListing: async (_, { listing }, { dataSources, userId, userRole }) => {
+      if (!userId) throw new AuthenticationError(authErrMessage);
+
+      const { title, description, photoThumbnail, numOfBeds, costPerNight, locationType, amenities } = listing;
+
+      if (userRole === 'Host') {
+        try {
+          const newListing = await dataSources.listingsAPI.createListing({
+            title,
+            description,
+            photoThumbnail,
+            numOfBeds,
+            costPerNight,
+            hostId: userId,
+            locationType,
+            amenities,
+          });
+
+          return {
+            code: 200,
+            success: true,
+            message: 'Listing successfully created!',
+            listing: newListing,
+          };
+        } catch (err) {
+          return {
+            code: 400,
+            success: false,
+            message: err.message,
+          };
+        }
+      } else {
+        return {
+          code: 400,
+          success: false,
+          message: 'Only hosts can create new listings',
+        };
+      }
+    },
+    updateListing: async (_, { listingId, listing }, { dataSources, userId }) => {
+      if (!userId) throw new AuthenticationError(authErrMessage);
+
+      try {
+        const updatedListing = await dataSources.listingsAPI.updateListing({ listingId, listing });
+
+        return {
+          code: 200,
+          success: true,
+          message: 'Listing successfully updated!',
+          listing: updatedListing,
         };
       } catch (err) {
         return {
@@ -158,6 +282,11 @@ const resolvers = {
       }
     },
   },
+  User: {
+    __resolveType(user) {
+      return user.role;
+    },
+  },
   Host: {
     overallRating: ({ id }, _, { dataSources }) => {
       return dataSources.reviewsDb.getOverallRatingForHost(id);
@@ -170,8 +299,8 @@ const resolvers = {
     },
   },
   Listing: {
-    host: ({ hostId }) => {
-      return { id: hostId };
+    host: ({ hostId }, _, { dataSources }) => {
+      return dataSources.accountsAPI.getUser(hostId);
     },
     overallRating: ({ id }, _, { dataSources }) => {
       return dataSources.reviewsDb.getOverallRatingForListing(id);
@@ -204,8 +333,8 @@ const resolvers = {
     checkOutDate: ({ checkOutDate }, _, { dataSources }) => {
       return dataSources.bookingsDb.getHumanReadableDate(checkOutDate);
     },
-    guest: ({ guestId }) => {
-      return { id: guestId };
+    guest: ({ guestId }, _, { dataSources }) => {
+      return dataSources.accountsAPI.getUser(guestId);
     },
     totalPrice: async ({ listingId, checkInDate, checkOutDate }, _, { dataSources }) => {
       const { totalCost } = await dataSources.listingsAPI.getTotalCost({ id: listingId, checkInDate, checkOutDate });
@@ -222,10 +351,14 @@ const resolvers = {
     },
   },
   Review: {
-    author: (review) => {
-      const role = review.targetType === 'LISTING' ? 'Guest' : 'Host';
-      return { __typename: role, id: review.authorId };
+    author: ({ authorId }, _, { dataSources }) => {
+      return dataSources.accountsAPI.getUser(authorId);
     },
+  },
+  AmenityCategory: {
+    ACCOMMODATION_DETAILS: 'Accommodation Details',
+    SPACE_SURVIVAL: 'Space Survival',
+    OUTDOORS: 'Outdoors',
   },
 };
