--- final/subgraph-listings/index.js
@@ -1,34 +1,0 @@
-const { ApolloServer, gql } = require('apollo-server');
-const { buildSubgraphSchema } = require('@apollo/subgraph');
-const { readFileSync } = require('fs');
-
-const typeDefs = gql(readFileSync('./schema.graphql', { encoding: 'utf-8' }));
-const resolvers = require('./resolvers');
-
-const ListingsAPI = require('./datasources/listings');
-const BookingsDb = require('./datasources/bookings');
-
-const server = new ApolloServer({
-  schema: buildSubgraphSchema({ typeDefs, resolvers }),
-  dataSources: () => {
-    return {
-      listingsAPI: new ListingsAPI(),
-      bookingsDb: new BookingsDb(),
-    };
-  },
-  context: ({ req }) => {
-    return { userId: req.headers.userid, userRole: req.headers.userrole };
-  },
-});
-
-const port = 4003;
-const subgraphName = 'listings';
-
-server
-  .listen({ port })
-  .then(({ url }) => {
-    console.log(`🚀 Subgraph ${subgraphName} running at ${url}`);
-  })
-  .catch((err) => {
-    console.error(err);
-  });
