--- final/subgraph-listings/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "listings-subgraph",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "nodemon index.js"
-  },
-  "nodemonConfig": {
-    "ext": "*.js, *.graphql"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@apollo/subgraph": "^0.1.1",
-    "apollo-server": "^3.4.0",
-    "graphql": "^15.6.1"
-  },
-  "devDependencies": {
-    "nodemon": "^2.0.13"
-  }
-}
