--- final/supergraph-config.yaml
@@ -1,13 +1,0 @@
-subgraphs:
-  monolith:
-    routing_url: http://localhost:4001
-    schema:
-      file: ./schema.graphql
-  accounts:
-    routing_url: http://localhost:4002
-    schema:
-      file: ./subgraph-accounts/schema.graphql
-  listings:
-    routing_url: http://localhost:4003
-    schema:
-      file: ./subgraph-listings/schema.graphql
